Time: 2516 ms (5.02%), Space: 6 MB (7.66%) - LeetHub

+++ 1523-count-odd-numbers-in-an-interval-range/1523-count-odd-numbers-in-an-interval-range.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int countOdds(int low, int high) {
-        int count=0;
+        long long int count=0;
         for(int i=low;i<=high;i++)
         {
             if(i%2!=0)
